E-greedy, Exp replay, entertainment
E-greedy is implemented in qmodel. A function is passed to it to determine how epsilon changes with the epoch. Experience replay was created, a deque that can gather and be sampled. The entertaining thing is the matplotlib func animation that will display and episode to keep you busy during training.